prometheus-operator: Add variable `watch_labeled_service_monitors`
This commit adds a variable `watch_labeled_service_monitors` to the
prometheus component config. Using this config user can decide if they
want to watch all the `ServiceMonitor` in the cluster or only those
labeled as `release: "prometheus-operator"`.

This commit sets the value of `serviceMonitorSelectorNilUsesHelmValues`
in the values file behind the scene to what user has provided.

When the value of `watch_labeled_service_monitors` is `true` the
`Prometheus` object's field `serviceMonitorSelector` is set with following:

```yaml
serviceMonitorSelector:
  matchLabels:
    release: "prometheus-operator"
```

Due to this any `ServiceMonitor` object that does not have above labels,
they won't be selected.

And when `serviceMonitorSelectorNilUsesHelmValues` is set to `false`,
`serviceMonitorSelector` field in `Prometheus` object is empty.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>