Makefile: add 'all' and 'build-test' targets
This commit adds 'build-test' target to Makefile, which will compile all
tests code, without running them, to make sure that syntax for all
tests, including e2e tests is correct. This is useful while refactoring.

It also adds 'all' target, which combines all possible checks for the
code into a single target to run, which should be run to run all
possible checks for the code.

Closes #93 #44

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>